Add Extension file, modified project
