+++ Navigation/SceneDelegate.swift
@@ -9,25 +9,6 @@ import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
-    func createFirstController() -> UINavigationController {
-        let navigationController = UINavigationController(rootViewController: FeedViewController())
-        navigationController.tabBarItem = UITabBarItem(title: "Feed", image: UIImage(systemName: "gamecontroller"), tag: 0)
-        return navigationController
-    }
-
-    func createSecondController() -> UINavigationController {
-        let navigationController = UINavigationController(rootViewController: ProfileViewController())
-        navigationController.tabBarItem = UITabBarItem(title: "Profile", image: UIImage(systemName: "figure.table.tennis"), tag: 1)
-        return navigationController
-    }
-
-    func createTabBarController() -> UITabBarController {
-        let tabBar = UITabBarController()
-        tabBar.viewControllers = [createFirstController(), createSecondController()]
-        tabBar.tabBar.backgroundColor = .lightGray
-        return tabBar
-    }
-
     var window: UIWindow?
 
 
@@ -40,35 +21,5 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window?.rootViewController = createTabBarController()
         window?.makeKeyAndVisible()
     }
-
-    func sceneDidDisconnect(_ scene: UIScene) {
-        // Called as the scene is being released by the system.
-        // This occurs shortly after the scene enters the background, or when its session is discarded.
-        // Release any resources associated with this scene that can be re-created the next time the scene connects.
-        // The scene may re-connect later, as its session was not necessarily discarded (see `application:didDiscardSceneSessions` instead).
-    }
-
-    func sceneDidBecomeActive(_ scene: UIScene) {
-        // Called when the scene has moved from an inactive state to an active state.
-        // Use this method to restart any tasks that were paused (or not yet started) when the scene was inactive.
-    }
-
-    func sceneWillResignActive(_ scene: UIScene) {
-        // Called when the scene will move from an active state to an inactive state.
-        // This may occur due to temporary interruptions (ex. an incoming phone call).
-    }
-
-    func sceneWillEnterForeground(_ scene: UIScene) {
-        // Called as the scene transitions from the background to the foreground.
-        // Use this method to undo the changes made on entering the background.
-    }
-
-    func sceneDidEnterBackground(_ scene: UIScene) {
-        // Called as the scene transitions from the foreground to the background.
-        // Use this method to save data, release shared resources, and store enough scene-specific state information
-        // to restore the scene back to its current state.
-    }
-
-
 }
